requirements.txt
scikit-image==0.22.0
torch==2.2.1
PyYAML==6.0.1
transformers==4.39.1
numpy==1.26.4
requests==2.32.3